requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask
flask-jwt-extended
flask-sqlalchemy
flask-cors
psycopg2-binary==2.9.6
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.24.3
opencv-python-headless==4.7.0.72
scipy==1.10.1
torch==2.0.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
bcrypt==3.2.0
Levenshtein
scikit-learn
Pillow